Изменены методы добавления, удаления и вывода информации клиентов в БД SQL

+++ src/savelying/Calculator.java
@@ -1,5 +1,5 @@
 package savelying;
 
 public interface Calculator <T extends Number> {
-	public double calculateFees (T clubId);
+	public int calculateFees (T clubId);
 }
